Start drawing finalized interventions in a normal layer that we control.
For the moment, disconnect DrawControls entirely, breaking most
functionality.
